LeetCode/tst/com/leetcode/problem222/countcompletetreenodes/SolutionTest.java
package com.leetcode.problem222.countcompletetreenodes;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

import com.leetcode.problem.utils.TreeNode;

/**
 * Given a complete binary tree, count the number of nodes.
 * 
 * Definition of a complete binary tree from Wikipedia: In a complete binary
 * tree every level, except possibly the last, is completely filled, and all
 * nodes in the last level are as far left as possible. It can have between 1
 * and 2h nodes inclusive at the last level h.
 * 
 * @author dongyuxi
 *
 */
public class SolutionTest {
    public Solution solution = null;

    @Before
    public void setup() {
        solution = new Solution();
    }

    @Test
    public void test() {
        TreeNode node11 = new TreeNode(1);
        TreeNode node21 = new TreeNode(2);
        TreeNode node22 = new TreeNode(7);
        TreeNode node31 = new TreeNode(1);
        TreeNode node32 = new TreeNode(3);
        TreeNode node33 = new TreeNode(6);
        TreeNode node34 = new TreeNode(9);
        node11.left = node21;
        node11.right = node22;
        node21.left = node31;
        node21.right = node32;
        node22.left = node33;
        node22.right = node34;

        assertEquals(7, solution.countNodes(node11));
    }

}
